requirements.txt
yfinance>=0.2.36
fredapi>=0.5.0
sqlalchemy>=2.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
apscheduler>=3.10.4
anthropic>=0.28.0
jsonschema>=4.22.0
python-dotenv>=1.0.1
